Extract duplicated format helpers to $lib/utils/format.ts
formatDuration, autoTitle, formatDate, formatDateLong, formatDateShort,
and formatTime were duplicated across 3 files. Now imported from a
single source.

Closes #20
