+++ src/lib/components/ActivityModal.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import { untrack } from 'svelte';
   import type { SessionSummary } from '$lib/tauri';
+  import { formatDuration, formatDate, autoTitle } from '$lib/utils/format';
 
   const ACTIVITY_TYPES = [
     { value: 'endurance', label: 'Endurance' },
@@ -44,27 +45,6 @@
   function handleClose() {
     dialogEl?.close();
     onClose();
-  }
-
-  function autoTitle(startTime: string): string {
-    const hour = new Date(startTime).getHours();
-    if (hour >= 5 && hour < 12) return 'Morning Ride';
-    if (hour >= 12 && hour < 17) return 'Afternoon Ride';
-    if (hour >= 17 && hour < 21) return 'Evening Ride';
-    return 'Night Ride';
-  }
-
-  function formatDuration(secs: number): string {
-    const h = Math.floor(secs / 3600);
-    const m = Math.floor((secs % 3600) / 60);
-    const s = secs % 60;
-    if (h > 0) return `${h}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`;
-    return `${m}:${String(s).padStart(2, '0')}`;
-  }
-
-  function formatDate(iso: string): string {
-    const d = new Date(iso);
-    return d.toLocaleDateString(undefined, { weekday: 'short', month: 'short', day: 'numeric' });
   }
 
   // Modal is freshly mounted each time — snapshot initial values (one-time read)
