+++ src/routes/history/+page.svelte
@@ -3,6 +3,7 @@
   import type { SessionSummary } from '$lib/tauri';
   import { api, extractError } from '$lib/tauri';
   import ActivityModal from '$lib/components/ActivityModal.svelte';
+  import { formatDuration, formatDateLong as formatDate, formatDateShort, formatTime, autoTitle } from '$lib/utils/format';
 
   let sessions = $state<SessionSummary[]>([]);
   let loading = $state(true);
@@ -38,14 +39,6 @@
     localStorage.setItem('historyView', mode);
   }
 
-  function autoTitle(startTime: string): string {
-    const hour = new Date(startTime).getHours();
-    if (hour >= 5 && hour < 12) return 'Morning Ride';
-    if (hour >= 12 && hour < 17) return 'Afternoon Ride';
-    if (hour >= 17 && hour < 21) return 'Evening Ride';
-    return 'Night Ride';
-  }
-
   function displayTitle(session: SessionSummary): string {
     return session.title ?? autoTitle(session.start_time);
   }
@@ -56,29 +49,6 @@
     tempo: 'Tempo', recovery: 'Recovery', race: 'Race', test: 'Test',
     warmup: 'Warmup', group_ride: 'Group Ride', free_ride: 'Free Ride', other: 'Other',
   };
-
-  function formatDate(iso: string): string {
-    const d = new Date(iso);
-    return d.toLocaleDateString(undefined, { weekday: 'short', month: 'short', day: 'numeric', year: 'numeric' });
-  }
-
-  function formatDateShort(iso: string): string {
-    const d = new Date(iso);
-    return d.toLocaleDateString(undefined, { month: 'short', day: 'numeric' });
-  }
-
-  function formatTime(iso: string): string {
-    const d = new Date(iso);
-    return d.toLocaleTimeString(undefined, { hour: '2-digit', minute: '2-digit' });
-  }
-
-  function formatDuration(secs: number): string {
-    const h = Math.floor(secs / 3600);
-    const m = Math.floor((secs % 3600) / 60);
-    const s = secs % 60;
-    if (h > 0) return `${h}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`;
-    return `${m}:${String(s).padStart(2, '0')}`;
-  }
 
   function rpeColor(value: number): string {
     if (value <= 5) {
